Don't depend on Action Pack from Railties
In 8ff310aef4e7ed424429059d10e3af0f169323fb, the Railties gem was split
from the Rails gem so that you could depend on it and skip some
frameworks, e.g. replace Active Record with something else.

Since then, Rails evolved and the current approach is to always depend
on Rails, but skipped frameworks are not required from
config/application.rb.

When writing a Railtie for Active Record or Active Job for example, it's
weird that the plugin ends up depending on Action Pack even if it's not
used at all.